+++ src/image/index.ios.ts
@@ -10,13 +10,16 @@ import {
     ImageInfo as ImageInfoBase,
     ImagePipelineConfigSetting,
     ScaleType,
+    SrcType,
     Stretch,
     failureImageUriProperty,
     imageRotationProperty,
     placeholderImageUriProperty,
     srcProperty,
-    stretchProperty
+    stretchProperty,
+    wrapNativeException
 } from './index-common';
+import { FailureEventData } from '@nativescript-community/ui-image';
 
 export class ImageInfo implements ImageInfoBase {
     constructor(
@@ -100,11 +103,52 @@ export function initialize(config?: ImagePipelineConfigSetting): void {
 }
 export function shutDown(): void {}
 
+function getContextFromOptions(options: Partial<Img>) {
+    const context: NSDictionary<string, any> = NSMutableDictionary.dictionary();
+    const transformers = [];
+    if (options.decodeWidth || options.decodeHeight) {
+        //@ts-ignore
+        transformers.push(NSImageDecodeSizeTransformer.transformerWithDecodeWidthDecodeHeight(options.decodeWidth || options.decodeHeight, options.decodeHeight || options.decodeWidth));
+    }
+    if (options.tintColor) {
+        transformers.push(SDImageTintTransformer.transformerWithColor(options.tintColor.ios));
+    }
+    if (options.blurRadius) {
+        transformers.push(SDImageBlurTransformer.transformerWithRadius(options.blurRadius));
+    }
+    if (options.roundAsCircle === true) {
+        //@ts-ignore
+        transformers.push(NSImageRoundAsCircleTransformer.transformer());
+    }
+    if (options.imageRotation !== 0 && !isNaN(options.imageRotation)) {
+        transformers.push(SDImageRotationTransformer.transformerWithAngleFitSize(-options.imageRotation * (Math.PI / 180), true));
+    }
+    if (options.roundBottomLeftRadius || options.roundBottomRightRadius || options.roundTopLeftRadius || options.roundTopRightRadius) {
+        transformers.push(
+            //@ts-ignore
+            NSImageRoundCornerTransformer.transformerWithTopLefRadiusTopRightRadiusBottomRightRadiusBottomLeftRadius(
+                layout.toDeviceIndependentPixels(options.roundTopLeftRadius),
+                layout.toDeviceIndependentPixels(options.roundTopRightRadius),
+                layout.toDeviceIndependentPixels(options.roundBottomRightRadius),
+                layout.toDeviceIndependentPixels(options.roundBottomLeftRadius)
+            )
+        );
+    }
+    if (options.mCIFilter) {
+        transformers.push(SDImageFilterTransformer.transformerWithFilter(options.mCIFilter));
+    }
+    if (transformers.length > 0) {
+        context.setValueForKey(SDImagePipelineTransformer.transformerWithTransformers(transformers), SDWebImageContextImageTransformer);
+    }
+    return context;
+}
+
 export class ImagePipeline {
     private mIos: SDImageCache = SDImageCache.sharedImageCache;
     constructor() {}
 
-    getCacheKey(uri: string, context) {
+    getCacheKey(uri: string, options: Partial<Img>) {
+        const context = getContextFromOptions(options);
         return SDWebImageManager.sharedManager.cacheKeyForURLContext(NSURL.URLWithString(uri), context);
     }
 
@@ -229,13 +273,13 @@ export class Img extends ImageBase {
     isLoading = false;
     mCacheKey: string;
 
-    contextOptions = null
+    contextOptions = null;
 
     get cacheKey() {
         return this.mCacheKey;
     }
     protected mImageSourceAffectsLayout: boolean = true;
-    protected mCIFilter: CIFilter;
+    mCIFilter: CIFilter;
     public createNativeView() {
         const result = this.animatedImageView ? SDAnimatedImageView.new() : UIImageView.new();
         result.contentMode = UIViewContentMode.ScaleAspectFit;
@@ -297,8 +341,9 @@ export class Img extends ImageBase {
     public async updateImageUri() {
         const imagePipeLine = getImagePipeline();
         const src = this.src;
-        if (!(src instanceof ImageSource)) {
-            const cachekKey = this.mCacheKey || getUri(src).absoluteString;
+        const srcType = typeof src;
+        if (src && (srcType === 'string' || src instanceof ImageAsset)) {
+            const cachekKey = this.mCacheKey || getUri(src as string | ImageAsset).absoluteString;
             // const isInCache = imagePipeLine.isInBitmapMemoryCache(cachekKey);
             // if (isInCache) {
             await imagePipeLine.evictFromCache(cachekKey);
@@ -359,8 +404,8 @@ export class Img extends ImageBase {
         if (error) {
             this.notify({
                 eventName: Img.failureEvent,
-                error
-            });
+                error: wrapNativeException(error)
+            } as FailureEventData);
             if (this.failureImageUri) {
                 image = this.getUIImage(this.failureImageUri);
                 this._setNativeImage(image, animate);
@@ -404,10 +449,21 @@ export class Img extends ImageBase {
     }
 
     protected async initImage() {
+        // this.nativeImageViewProtected.setImageURI(null);
+        this.handleImageSrc(this.src);
+    }
+
+    protected async handleImageSrc(src: SrcType) {
         if (this.nativeViewProtected) {
-            const src = this.src;
             if (src instanceof Promise) {
-                this.src = await src;
+                this.handleImageSrc(await src);
+                return;
+            } else if (typeof src === 'function') {
+                const newSrc = src();
+                if (newSrc instanceof Promise) {
+                    await newSrc;
+                }
+                this.handleImageSrc(newSrc);
                 return;
             }
             if (src) {
@@ -428,7 +484,7 @@ export class Img extends ImageBase {
                     }
                 }
 
-                const uri = getUri(src);
+                const uri = getUri(src as string | ImageAsset);
                 this.isLoading = true;
                 let options = SDWebImageOptions.ScaleDownLargeImages | SDWebImageOptions.AvoidAutoSetImage;
 
@@ -436,7 +492,7 @@ export class Img extends ImageBase {
                     this.placeholderImage = this.getUIImage(this.placeholderImageUri);
                     this._setNativeImage(this.placeholderImage, animate);
                 }
-                
+
                 if (this.noCache) {
                     // const key = uri.absoluteString;
                     // const imagePipeLine = getImagePipeline();
@@ -449,55 +505,20 @@ export class Img extends ImageBase {
                 if (this.alwaysFade === true) {
                     options |= SDWebImageOptions.ForceTransition;
                 }
-                const context: NSDictionary<string, any> = NSMutableDictionary.dictionary();
-                const transformers = [];
                 if (this.progressiveRenderingEnabled === true) {
                     options = options | SDWebImageOptions.ProgressiveLoad;
                 }
-                if (this.decodeWidth || this.decodeHeight) {
-                    //@ts-ignore
-                    transformers.push(NSImageDecodeSizeTransformer.transformerWithDecodeWidthDecodeHeight(this.decodeWidth || this.decodeHeight, this.decodeHeight || this.decodeWidth));
+                const context = getContextFromOptions(this);
+                if (this.animatedImageView) {
+                    // as we use SDAnimatedImageView  all images are loaded as SDAnimatedImage;
+                    options |= SDWebImageOptions.TransformAnimatedImage;
                 }
-                if (this.tintColor) {
-                    transformers.push(SDImageTintTransformer.transformerWithColor(this.tintColor.ios));
-                }
-                if (this.blurRadius) {
-                    transformers.push(SDImageBlurTransformer.transformerWithRadius(this.blurRadius));
-                }
-                if (this.roundAsCircle === true) {
-                    //@ts-ignore
-                    transformers.push(NSImageRoundAsCircleTransformer.transformer());
-                }
-                if (this.imageRotation !== 0 && !isNaN(this.imageRotation)) {
-                    transformers.push(SDImageRotationTransformer.transformerWithAngleFitSize(-this.imageRotation * (Math.PI / 180), true));
-                }
-                if (this.roundBottomLeftRadius || this.roundBottomRightRadius || this.roundTopLeftRadius || this.roundTopRightRadius) {
-                    transformers.push(
-                        //@ts-ignore
-                        NSImageRoundCornerTransformer.transformerWithTopLefRadiusTopRightRadiusBottomRightRadiusBottomLeftRadius(
-                            layout.toDeviceIndependentPixels(this.roundTopLeftRadius),
-                            layout.toDeviceIndependentPixels(this.roundTopRightRadius),
-                            layout.toDeviceIndependentPixels(this.roundBottomRightRadius),
-                            layout.toDeviceIndependentPixels(this.roundBottomLeftRadius)
-                        )
-                    );
-                }
-                if (this.mCIFilter) {
-                    transformers.push(SDImageFilterTransformer.transformerWithFilter(this.mCIFilter));
-                }
-                if (transformers.length > 0) {
-                    if (this.animatedImageView) {
-                        // as we use SDAnimatedImageView  all images are loaded as SDAnimatedImage;
-                        options |= SDWebImageOptions.TransformAnimatedImage;
-                    }
-                    context.setValueForKey(SDImagePipelineTransformer.transformerWithTransformers(transformers), SDWebImageContextImageTransformer);
-                    // context.setValueForKey(SDImageCacheType.Memory, SDWebImageContextOriginalStoreCacheType);
-                }
-                if (this.contextOptions && typeof this.contextOptions === 'object') {                    
-                    Object.keys(this.contextOptions).forEach(k=>{
-                        let value = this.contextOptions[k];
+
+                if (this.contextOptions && typeof this.contextOptions === 'object') {
+                    Object.keys(this.contextOptions).forEach((k) => {
+                        const value = this.contextOptions[k];
                         context.setValueForKey(value, k);
-                    })
+                    });
                 }
                 this.mCacheKey = SDWebImageManager.sharedManager.cacheKeyForURLContext(uri, context);
                 this.nativeImageViewProtected.sd_setImageWithURLPlaceholderImageOptionsContextProgressCompleted(